Refactor toolsPanel into separate module
Prep for adding in a profile module that uses the same code.
